29/08 Finalizando atividade

+++ primeiro_projeto/src/main/java/jv/triersistemas/primeiro_projeto/service/impl/CategoriaServiceImpl.java
@@ -1,15 +1,16 @@
 package jv.triersistemas.primeiro_projeto.service.impl;
 
-import jv.triersistemas.primeiro_projeto.dto.CategoriaDto;
-import jv.triersistemas.primeiro_projeto.entity.CategoriaEntity;
-import jv.triersistemas.primeiro_projeto.repository.CategoriaRepository;
-import jv.triersistemas.primeiro_projeto.service.CategoriaService;
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import jv.triersistemas.primeiro_projeto.dto.CategoriaDto;
+import jv.triersistemas.primeiro_projeto.entity.CategoriaEntity;
+import jv.triersistemas.primeiro_projeto.repository.CategoriaRepository;
+import jv.triersistemas.primeiro_projeto.service.CategoriaService;
 
 @Primary
 @Service
@@ -41,22 +42,33 @@ public class CategoriaServiceImpl implements CategoriaService {
 	}
 
 	@Override
-	public CategoriaDto atualizarCategoria(Long id, CategoriaDto categoriaDto) {
+	public CategoriaDto atualizarCategoria(Long id, CategoriaDto categoriaDto) throws RuntimeException {
 
-		var categoriaEntity = buscaIdBanco(id);
+		var categoriaEntityOptional = buscaIdBanco(id);
 
-		if (categoriaEntity.isPresent()) {
-			categoriaRepository.save(categoriaEntity.get().atualizaRegistro(categoriaDto));
+		if (categoriaEntityOptional.isPresent()) {
 
-			return new CategoriaDto(categoriaEntity.get());
+			var categoriaEntity = categoriaEntityOptional.get().atualizaRegistro(categoriaDto);
+
+			categoriaRepository.save(categoriaEntity);
+
+			return new CategoriaDto(categoriaEntity);
 		}
 
 		return null;
 	}
 
 	@Override
-	public void deletarCategoria(Long id) {
-		categoriaRepository.deleteById(id);
+	public void deletarCategoria(Long id) throws RuntimeException {
+		var categoriaEntity = categoriaRepository.findById(id);
+
+		if (categoriaEntity.isPresent()) {
+			if (categoriaEntity.get().getTarefas().isEmpty()) {
+				categoriaRepository.deleteById(id);
+			}
+			throw new IllegalArgumentException(
+					"ERRO: Delete primeiro as tarefas que possuem essa categoria antes de exclui-lá");
+		}
 
 	}
 
@@ -64,4 +76,5 @@ public class CategoriaServiceImpl implements CategoriaService {
 	public Optional<CategoriaEntity> buscaIdBanco(Long id) {
 		return categoriaRepository.findById(id);
 	}
+
 }
